New simulation

+++ project/route.py
@@ -9,9 +9,9 @@ def route(graph, locations):
     return path
 
 def plot_route(nodes, locations, path, distance, filename):
-    m = folium.Map(location = nodes[int(path[math.ceil(len(path) / 2)])])
-    path_coords = [nodes[int(index)] for index in path]
+    m = folium.Map(location = nodes[path[math.ceil(len(path) / 2)]])
+    path_coords = [nodes[index] for index in path]
     folium.PolyLine(path_coords, color="red", weight=4.5).add_to(m)
     for i in range(len(locations)):
-        folium.Marker(nodes[int(locations[i])], popup=f"The total route distance is {round(distance / 1000, 3)}km.").add_to(m)
+        folium.Marker(nodes[locations[i]], popup=f"The total route distance is {round(distance / 1000, 3)}km.").add_to(m)
     m.save(f"maps/{filename}")
